+++ cucumber-compatibility-kit/javascript/features/stack-traces/stack-traces.ts
@@ -1,0 +1,5 @@
+import { When } from 'fake-cucumber'
+
+When('a step throws an exception', function() {
+  throw new Exception('BOOM')
+})