+++ cucumber-react/javascript/src/CucumberQueryContext.ts
@@ -1,4 +1,5 @@
 import React from 'react'
 import CucumberQuery from 'cucumber-query'
+import { GherkinQuery } from 'gherkin'
 
-export default React.createContext(new CucumberQuery())
+export default React.createContext(new CucumberQuery(new GherkinQuery()))