+++ fake-cucumber/javascript/src/makeTestCase.ts
@@ -1,4 +1,5 @@
 import { IdGenerator, messages } from 'cucumber-messages'
+import { GherkinQuery } from 'gherkin'
 import TestCase from './TestCase'
 import IStepDefinition from './IStepDefinition'
 import IHook from './IHook'
@@ -11,13 +12,34 @@ export default function makeTestCase(
   stepDefinitions: IStepDefinition[],
   beforeHooks: IHook[],
   afterHooks: IHook[],
+  gherkinQuery: GherkinQuery,
   newId: IdGenerator.NewId
 ): TestCase {
-  const beforeHookSteps = makeHookSteps(pickle, beforeHooks, false, newId)
-  const pickleTestSteps = pickle.steps.map(pickleStep =>
-    makePickleTestStep(newId(), pickleStep, stepDefinitions)
+  const beforeHookSteps = makeHookSteps(
+    pickle,
+    beforeHooks,
+    false,
+    gherkinQuery,
+    newId
   )
-  const afterHookSteps = makeHookSteps(pickle, afterHooks, true, newId)
+  const pickleTestSteps = pickle.steps.map(pickleStep => {
+    const sourceFrames = pickleStep.astNodeIds.map(
+      astNodeId => `${pickle.uri}:${gherkinQuery.getLocation(astNodeId).line}`
+    )
+    return makePickleTestStep(
+      newId(),
+      pickleStep,
+      stepDefinitions,
+      sourceFrames
+    )
+  })
+  const afterHookSteps = makeHookSteps(
+    pickle,
+    afterHooks,
+    true,
+    gherkinQuery,
+    newId
+  )
   const testSteps: ITestStep[] = []
     .concat(beforeHookSteps)
     .concat(pickleTestSteps)
@@ -30,6 +52,7 @@ function makeHookSteps(
   pickle: messages.IPickle,
   hooks: IHook[],
   alwaysExecute: boolean,
+  gherkinQuery: GherkinQuery,
   newId: IdGenerator.NewId
 ): ITestStep[] {
   return hooks
@@ -37,9 +60,18 @@ function makeHookSteps(
       const supportCodeExecutor = hook.match(pickle)
       if (supportCodeExecutor !== null) {
         const id = newId()
-        return new HookTestStep(id, hook.id, alwaysExecute, [
-          supportCodeExecutor,
-        ])
+
+        const sourceFrames = pickle.astNodeIds.map(
+          astNodeId =>
+            `${pickle.uri}:${gherkinQuery.getLocation(astNodeId).line}`
+        )
+        return new HookTestStep(
+          id,
+          hook.id,
+          alwaysExecute,
+          [supportCodeExecutor],
+          sourceFrames
+        )
       }
     })
     .filter(testStep => testStep !== undefined)