+++ fake-cucumber/javascript/test/CucumberTest.ts
@@ -6,6 +6,7 @@ import assert from 'assert'
 import SupportCode from '../src/SupportCode'
 import makeDummyStepDefinitions from './makeDummyStepDefinitions'
 import makeDummyHooks from './makeDummyHooks'
+import { GherkinQuery } from 'gherkin'
 
 describe('Cucumber', () => {
   it('runs tagged hooks', async () => {
@@ -16,9 +17,13 @@ describe('Cucumber', () => {
     Given a passed step
   `
 
+    const gherkinQuery = new GherkinQuery()
     const gherkinMessageList = await streamToArray(
       gherkinMessages(feature, 'test.feature')
     )
+    for (const envelope of gherkinMessageList) {
+      gherkinQuery.update(envelope)
+    }
     const supportCode = new SupportCode(IdGenerator.incrementing())
     makeDummyStepDefinitions(supportCode)
     makeDummyHooks(supportCode)
@@ -27,6 +32,7 @@ describe('Cucumber', () => {
       supportCode.stepDefinitions,
       supportCode.beforeHooks,
       supportCode.afterHooks,
+      gherkinQuery,
       IdGenerator.incrementing()
     )
     const messageList: messages.IEnvelope[] = []